Mov. Config* a. related Classes to their own files
